Fix review demo

+++ KNSemiModalViewControllerDemo/KNTableDemoController.m
@@ -38,7 +38,7 @@
   if (!cell) {
     cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:CellIdentifier];
   }
-  cell.textLabel.text = [NSString stringWithFormat:@"Demo row %d", indexPath.row];
+  cell.textLabel.text = [NSString stringWithFormat:@"Demo row %ld", (long)indexPath.row];
   return cell;
 }
